renamed Mock Transaction.py, added more indentation, and richer formating
